requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
requests>=2.31.0
jinja2>=3.1.2
sqlalchemy>=2.0.23
pyJWT>=2.8.0
Pillow>=10.1.0
passlib[bcrypt]>=1.7.4
aiofiles>=23.2.1
python-multipart>=0.0.6
langchain>=0.0.350
langchain-huggingface>=0.0.3
langchain-community>=0.0.10
faiss-cpu>=1.7.4
torch>=2.1.1
opencv-python>=4.8.1
transformers>=4.35.2
sentence-transformers>=2.2.2
huggingface-hub>=0.19.4
gTTS>=2.4.0
faster-whisper>=0.10.0
openai-whisper>=20231117
sentencepiece>=0.1.99
pydantic>=2.5.0
python-dotenv>=1.0.0
google-generativeai>=0.3.2
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2
websockets>=12.0
python-bcrypt>=3.2.2

